requirements.txt
asgiref==3.11.0
Django==6.0.1
django-jet-reboot==1.3.10
django-semantic-forms==0.1.8
django-tinymce==5.0.0
django-widget-tweaks==1.5.1
google-api-python-client==1.4.1
httplib2==0.31.0
oauth2client==4.1.3
pillow==12.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.3.1
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
uritemplate==4.2.0
mysqlclient==2.2.7